+++ packages/frontend/src/core/geo/map/MapWrapper.ts
@@ -288,24 +288,12 @@ export class MapWrapper {
   }
 
   public getTextAttributions(): string[] {
-    const div = document.createElement('div');
     return _(this.getLayers())
       .filter((lay) => lay.isVisible())
-      .flatMap((lay) => lay.getSource().getAttributions())
-      .flatMap((attr) => {
-        if (!attr) {
-          return [];
-        }
-
-        const attributions = attr({} as any); // Typings are borked
-        return attributions instanceof Array ? attributions : [attributions];
-      })
+      .flatMap((lay) => lay.getAttributions())
       .uniq() // Some attributions can appear twice
-      .map((attr) => {
-        div.innerHTML = attr;
-        return div.textContent || div.innerText || '';
-      })
-      .value();
+      .filter((attr) => !!attr)
+      .value() as string[];
   }
 
   public setView(view: AbcView) {